+++ CHANGELOG.md
@@ -5,7 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v0.2.0] - 2024-01-31
+## [v0.3.1](https://github.com/justDeevin/NuhxBoard/releases/v0.3.1) - 2024-2-2
+
+### Fixed
+
+- Window size under Wayland. [6af88ac](https://github.com/justDeevin/NuhxBoard/commit/6af88ac0ec2c390a60ce4f1a2648284dd271be9c)
+- Occasional duration_since error while window is focused. [f0ef286](https://github.com/justDeevin/NuhxBoard/commit/f0ef286c50a18dec68cf8395c97c20f97799356a)
+
+## [v0.3.0](https://github.com/justDeevin/NuhxBoard/releases/v0.3.0) - 2024-2-1
+
+### Added
+
+- Function to list keyboards and keyboard groups. [67e5308](https://github.com/justDeevin/NuhxBoard/commit/67e5308a3b7b76253bef1b7de5dd8d830190d35c)
+- Warning when running under Wayland. [b389725](https://github.com/justDeevin/NuhxBoard/commit/b3897255979f55006802939eee9dab4bcc03c478)
+
+### Changed
+
+- Help message tagline now matches NuhxBoard's cross-platform capability. [d300ee9](https://github.com/justDeevin/NuhxBoard/commit/d300ee9f8902d8c745b47662c9319061c317b7e7)
+
+### Fixed
+
+- Occasional panic on window close. [aad1d09](https://github.com/justDeevin/NuhxBoard/commit/aad1d0997be01f58092feb43fcc81625b717c450)
+- Events not being handled when NuhxBoard is focused. [#4](https://github.com/justDeevin/NuhxBoard/issues/4)
+
+## [v0.2.1](https://github.com/thepyrotf2/nuhxboard/releases/v0.2.1) - 2024-01-31
 
 ### Added